refactor: optional 수정
